requirements.txt
Flask==2.0.1
tinydb==4.5.1
PyYAML==5.4.1
waitress==2.0.0

# For Image Files
Pillow==8.3.1

# For Video Files
moviepy==1.0.3

# CLI
python-magic==0.4.27